requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.3.0
matplotlib>=3.4.0
Pillow>=8.0.0